update kalosm-language to new task interface
