+++ interfaces/kalosm-language/src/search/preprocessing/hypothetical.rs
@@ -1,8 +1,10 @@
-use kalosm_language_model::Embedder;
-use kalosm_sample::{LiteralParser, ParserExt, StopOn};
+use std::borrow::Cow;
+
+use kalosm_language_model::{CreateChatSession, Embedder, StructuredChatModel};
+use kalosm_sample::{CreateParserState, LiteralParser, ParseResult, ParseStatus, Parser, ParserExt, StopOn};
 
 use crate::{
-    prelude::{Document, IndexParser, StructuredRunner, Task},
+    prelude::{Document, Task},
     search::Chunk,
 };
 
@@ -43,13 +45,14 @@ fn create_constraints() -> Constraints {
 }
 
 /// A builder for a hypothetical chunker.
-pub struct HypotheticalBuilder {
+pub struct HypotheticalBuilder<M: CreateChatSession> {
+    model: M,
     task_description: Option<String>,
     examples: Option<Vec<(String, String)>>,
     chunking: Option<ChunkStrategy>,
 }
 
-impl HypotheticalBuilder {
+impl<M: CreateChatSession> HypotheticalBuilder<M> {
     /// Set the chunking strategy.
     pub fn with_chunking(mut self, chunking: ChunkStrategy) -> Self {
         self.chunking = Some(chunking);
@@ -77,7 +80,7 @@ impl HypotheticalBuilder {
     }
 
     /// Build the hypothetical chunker.
-    pub fn build(self) -> Hypothetical {
+    pub fn build(self) -> Hypothetical<M> {
         let task_description = self
             .task_description
             .unwrap_or_else(|| TASK_DESCRIPTION.to_string());
@@ -89,25 +92,23 @@ impl HypotheticalBuilder {
         });
         let chunking = self.chunking;
 
-        let task = Task::builder(task_description)
-            .with_constraints(create_constraints())
-            .with_examples(examples)
-            .build();
+        let task = Task::new(self.model, task_description).with_examples(examples);
 
         Hypothetical { chunking, task }
     }
 }
 
 /// Generates questions for a document.
-pub struct Hypothetical {
+pub struct Hypothetical<M: CreateChatSession> {
     chunking: Option<ChunkStrategy>,
-    task: Task<StructuredRunner<Constraints>>,
+    task: Task<M>,
 }
 
-impl Hypothetical {
+impl<M: CreateChatSession> Hypothetical<M> {
     /// Create a new hypothetical generator.
-    pub fn builder() -> HypotheticalBuilder {
+    pub fn builder(model: M) -> HypotheticalBuilder<M> {
         HypotheticalBuilder {
+            model,
             task_description: None,
             examples: None,
             chunking: None,
@@ -115,13 +116,17 @@ impl Hypothetical {
     }
 
     /// Generate a list of hypothetical questions about the given text.
-    pub async fn generate_question<M>(&self, text: &str, model: &M) -> Result<Vec<String>, M::Error>
+    pub async fn generate_question(&self, text: &str) -> Result<Vec<String>, M::Error>
     where
-        M: Model,
-        <M::SyncModel as SyncModel>::Session: Sync + Send,
-        M::Error: std::fmt::Debug,
+        M: StructuredChatModel<Constraints> + Send + Sync + Clone + Unpin + 'static,
+        M::ChatSession: Clone + Send + Sync + Unpin + 'static,
+        M::Error: Send + Sync + Unpin,
     {
-        let questions = self.task.run(text, model).result().await?;
+        let questions = self
+            .task
+            .run(text)
+            .with_constraints(create_constraints())
+            .await?;
         let documents = questions
             .1
             .into_iter()
@@ -130,18 +135,6 @@ impl Hypothetical {
 
         Ok(documents)
     }
-
-    /// Turn this hypothetical generator into a chunker.
-    pub fn chunker<'a, M>(&'a self, model: &'a M) -> HypotheticalChunker<'a, M>
-    where
-        M: Model,
-        <M::SyncModel as SyncModel>::Session: Sync + Send,
-    {
-        HypotheticalChunker {
-            hypothetical: self,
-            model,
-        }
-    }
 }
 
 /// An error that can occur when chunking a document with [`HypotheticalChunker`].
@@ -149,26 +142,17 @@ impl Hypothetical {
 pub enum HypotheticalChunkerError<E1: Send + Sync + 'static, E2: Send + Sync + 'static> {
     /// An error from the text generation model.
     #[error("Text generation model error: {0}")]
-    TextModelError(#[from] StructuredTextGenerationError<E1>),
+    TextModelError(#[from] E1),
     /// An error from the embedding model.
     #[error("Embedding model error: {0}")]
     EmbeddingModelError(E2),
 }
 
-/// A hypothetical chunker.
-pub struct HypotheticalChunker<'a, M: Model>
+impl<M> Chunker for Hypothetical<M>
 where
-    <M::SyncModel as SyncModel>::Session: Sync + Send,
-{
-    hypothetical: &'a Hypothetical,
-    model: &'a M,
-}
-
-impl<'a, M> Chunker for HypotheticalChunker<'a, M>
-where
-    M: Model,
-    <M::SyncModel as SyncModel>::Session: Sync + Send,
-    M::Error: std::fmt::Debug,
+    M: StructuredChatModel<Constraints> + Send + Sync + Clone + Unpin + 'static,
+    M::ChatSession: Clone + Send + Sync + Unpin + 'static,
+    M::Error: Send + Sync + Unpin,
 {
     type Error<E: Send + Sync + 'static> = HypotheticalChunkerError<M::Error, E>;
 
@@ -181,7 +165,6 @@ where
 
         #[allow(clippy::single_range_in_vec_init)]
         let byte_chunks = self
-            .hypothetical
             .chunking
             .map(|chunking| chunking.chunk_str(body))
             .unwrap_or_else(|| vec![0..body.len()]);
@@ -194,10 +177,7 @@ where
         let mut questions_count = Vec::new();
         for byte_chunk in &byte_chunks {
             let text = &body[byte_chunk.clone()];
-            let mut chunk_questions = self
-                .hypothetical
-                .generate_question(text, self.model)
-                .await?;
+            let mut chunk_questions = self.generate_question(text).await?;
             questions.append(&mut chunk_questions);
             questions_count.push(chunk_questions.len());
         }
@@ -228,5 +208,119 @@ where
         }
 
         Ok(chunks)
+    }
+}
+
+
+/// The state of the [`IndexParser`] parser
+#[derive(Debug, Clone)]
+pub (crate)struct IndexParserState<PA> {
+    states: Vec<ParseResult<PA>>,
+}
+
+/// A parser that parses a sequence of parsers and returns the index of the first parser that succeeds
+#[derive(Debug, Clone)]
+pub(crate) struct IndexParser<S: Parser> {
+    parsers: Vec<S>,
+}
+
+impl<S: Parser> IndexParser<S> {
+    /// Create a new index parser
+    pub fn new(parsers: Vec<S>) -> Self {
+        Self { parsers }
+    }
+}
+
+impl<S: CreateParserState> CreateParserState for IndexParser<S> {
+    fn create_parser_state(&self) -> Self::PartialState {
+        IndexParserState {
+            states: self
+                .parsers
+                .iter()
+                .map(|s| Ok(s.create_parser_state()))
+                .collect(),
+        }
+    }
+}
+
+impl<S: Parser> Parser for IndexParser<S> {
+    type Output = (usize, S::Output);
+    type PartialState = IndexParserState<S::PartialState>;
+
+    fn parse<'a>(
+        &self,
+        state: &Self::PartialState,
+        input: &'a [u8],
+    ) -> ParseResult<kalosm_sample::ParseStatus<'a, Self::PartialState, Self::Output>> {
+        let mut states = state.states.clone();
+        let mut has_incomplete_option = false;
+        let mut required_next: Option<Cow<'static, str>> = None;
+        let last_index = self.parsers.len() - 1;
+        for (i, parser) in self.parsers.iter().enumerate() {
+            match &states[i] {
+                Ok(state) => {
+                    let result = parser.parse(state, input);
+                    match result {
+                        Ok(ParseStatus::Finished {
+                            result,
+                            remaining: r,
+                        }) => {
+                            return Ok(ParseStatus::Finished {
+                                result: (i, result),
+                                remaining: r,
+                            })
+                        }
+                        Ok(ParseStatus::Incomplete {
+                            new_state: s,
+                            required_next: new_required_next,
+                        }) => {
+                            states[i] = Ok(s);
+                            has_incomplete_option = true;
+                            match required_next {
+                                Some(r) => {
+                                    let mut common_bytes = 0;
+                                    for (byte1, byte2) in r.bytes().zip(new_required_next.bytes()) {
+                                        if byte1 != byte2 {
+                                            break;
+                                        }
+                                        common_bytes += 1;
+                                    }
+                                    required_next = Some(match (r, new_required_next) {
+                                        (Cow::Borrowed(required_next), _) => {
+                                            Cow::Borrowed(&required_next[common_bytes..])
+                                        }
+                                        (_, Cow::Borrowed(required_next)) => {
+                                            Cow::Borrowed(&required_next[common_bytes..])
+                                        }
+                                        (Cow::Owned(mut required_next), _) => {
+                                            required_next.truncate(common_bytes);
+                                            Cow::Owned(required_next)
+                                        }
+                                    });
+                                }
+                                None => {
+                                    required_next = Some(new_required_next);
+                                }
+                            }
+                        }
+                        Err(e) => {
+                            if !has_incomplete_option && i == last_index {
+                                return Err(e);
+                            }
+                            states[i] = Err(e);
+                        }
+                    }
+                }
+                Err(err) => {
+                    if !has_incomplete_option && i == last_index {
+                        return Err(err.clone());
+                    }
+                }
+            }
+        }
+        Ok(ParseStatus::Incomplete {
+            new_state: IndexParserState { states },
+            required_next: required_next.unwrap_or_default(),
+        })
     }
 }
